+++ src/lib/quotes.ts
