+++ src/lib/timer.ts
